+++ Data/BodyFormulaData.cs
@@ -16,7 +16,10 @@ namespace CUCoreLib.Data
         public Dictionary<string, float> JumpSpeed = new Dictionary<string, float>();
         public Dictionary<string, float> AveragePain = new Dictionary<string, float>();
 
+        [JsonIgnore]
         public float AppliedJumpSpeedContribution;
+
+        [JsonIgnore]
         public float AppliedAveragePainContribution;
 
         [JsonIgnore]